Change data type of Population to long long int

+++ KMITL_C-CPP/Semester_2/Lab/Lab4/Region.cpp
@@ -1,7 +1,7 @@
 #include "Region.h"
 
 Region::Region() {}
-Region::Region(string key, string name, double latitude, double longitude, long int population, double area) {
+Region::Region(string key, string name, double latitude, double longitude, long long int population, double area) {
     this->key = key;
     this->name = name;
     this->latitude = latitude;
@@ -14,12 +14,12 @@ string Region::getKey() { return this->key; }
 string Region::getName() { return this->name; }
 double Region::getLatitude() { return this->latitude; }
 double Region::getLongitude() { return this->longitude; }
-double Region::getPopulation() { return this->population; }
+long long int Region::getPopulation() { return this->population; }
 double Region::getArea() { return this->area; }
 double Region::getPopulationDensity() { return this->population / this->area; }
 
 Province::Province() {}
-Province::Province(string key, string name, double latitude, double longitude, long int population, double area) : Region(key, name, latitude, longitude, population, area) {}
+Province::Province(string key, string name, double latitude, double longitude, long long int population, double area) : Region(key, name, latitude, longitude, population, area) {}
 
 Sub_District::Sub_District() {}
-Sub_District::Sub_District(string key, string name, double latitude, double longitude, long int population, double area) : Region(key, name, latitude, longitude, population, area) {}
+Sub_District::Sub_District(string key, string name, double latitude, double longitude, long long int population, double area) : Region(key, name, latitude, longitude, population, area) {}
